修正血量校正以及開場UI

+++ Assets/Script/Player/PlayerController.cs
@@ -192,7 +192,7 @@ public class PlayerController : GenericEntityController {
 				this.enabled = false;
 			}
 
-			float hpProportion = hp / mostHp;
+			float hpProportion = (float)hp / mostHp;
 
 			HpController.SetValue (hpProportion);
 		}
